Allow empty path in endpoint config
Similar to the default endpoint configurations for list and create, this will allow users to specify an empty path param resulting in the basePath being the full url for that endpoint. An example use case for this might be to create a request that acts the same as 'list', but is invoked under a different name.